Don't make the caller do work the function can do
